# melpaBuild: allow nix unstable version format