@AndersonTorres 2024-06-02
I believe this is an extra reason to merge https://github.com/NixOS/nixpkgs/pull/315949 :)