@jian-lin 2024-06-11
The [current PR](https://github.com/NixOS/nixpkgs/pull/316075) to update staging-next takes a bit more time than usual.